Tidy rev Q1 actually wasn't fixed; remove for now; join issue
Full joins should basically always be performed on a unique key for each of the
sides unless we are willing to do complex custom logic to fix them up. Otherwise
we get single rows invented for what normally should be many. It only looks like
it's doing something somewhat reasonable because we only have a single time
value. But we'd still get just a single new row per missing time series geo even
if there were multiple time values; that shouldn't be filled in with a single
`time_value`.