Fix revision visualization
Previously, the `snapshots` variable in `ggplot(snapshots) `is undefined. I suspect the reason quarto still compiles is because of the cached computation results.

This committ addresses this. I changed `epi_archive` example to use `archive_cases_dv_subset_all_states`  and plot revision patterns for CA and UT, which is consistent with previous examples.